+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 - Add `Pagination` class to work with pagination
 - Add `Repository` interface with common repository methods
 - Add `Timestamp` value class for representing timestamps
+- Add `Latitude` and `Longitude` value classes for representing geographic coordinates
+- Add `PlaceId` value class for representing place identifiers on Google Maps and Apple Maps
+- Add `CustomGeolocation` value class for representing custom geographic locations
 `core-mit` module
 - Add `DetailsConfiguration` for `WrappedException.kt` and `EncodableResult.kt` to customize exception details inclusion
 - Add `randomBase64()` method to generate random Base64 of the specified length